sd_server/requirements.txt
django==1.7.4
argparse==1.2.1
django-configurations==0.8
